==> notebooks/04_transform_raw_data_into_features_and_targets.ipynb <==
## Tranforming raw data into features and targets

!pwd

import os
os.chdir('..')

# to reload the kernel with latest changes that is made to .py files
%reload_ext autoreload
%autoreload 2

from src.data import load_raw_data

rides = load_raw_data(year=2022)
rides

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)
ts_data

from src.data import transform_ts_data_into_features_and_target

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len=24*28*1, # one month of history
    step_size=24,
)

print(f"{features.shape=}")

print(f"{targets.shape=}")

final_data = features
final_data['target_rides_next_hour'] = targets

from src.paths import TRANSFORMED_DATA_DIR
final_data.to_parquet(TRANSFORMED_DATA_DIR/ 'final_data.parquet')
