==> notebooks/05_visualize_training_data.ipynb <==
!pwd

import os
os.chdir('..')

%reload_ext autoreload
%autoreload 2

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

feature_and_target = pd.read_parquet(TRANSFORMED_DATA_DIR / 'final_data.parquet')
feature_and_target

features = feature_and_target.drop(columns=['target_rides_next_hour'])
targets = feature_and_target['target_rides_next_hour']

from src.plot import plot_one_sample

# plotting the first row
plot_one_sample(
    features,
    targets,
    example_id=0,
)

plot_one_sample(
    features,
    targets,
    example_id=5,
)
