==> notebooks/06_baseline_model.ipynb <==
!pwd

import os
os.chdir("..")

%reload_ext autoreload
%autoreload 2

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / "final_data.parquet")
df

df.dtypes

from src.data_split import train_test_split
from datetime import datetime
import pytz

cutoff_date = datetime(2022, 8, 1, 0, 0, 0)

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=cutoff_date,
    target_column_name='target_rides_next_hour'
)

print(f"{X_train.shape=}")
print(f"{y_train.shape=}")
print(f"{X_test.shape=}")
print(f"{y_test.shape=}")

## Creating baseline model 1: predicting previous 1 hour rides

#### Inroder to predict taxi demand for the next hour, we'll look at what was the demand for previous hour

import numpy as np

class BaselineModelPreviousHour:
    """
    Prediction = actual demand observed in the last hour
    """

    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass

    def predict(self, X_test: pd.DataFrame) -> np.array:
        """"""
        return X_test[f'rides_previous_1_hour']

model = BaselineModelPreviousHour() # created model object
predictions = model.predict(X_test)
predictions

from sklearn.metrics import mean_absolute_error

test_mae = mean_absolute_error(y_test, predictions)
print(f"{test_mae=:.4f}")


## Creating baseline model 2: predicting previous 1 week rides

class BaselineModelPreviousWeek:
    """
    Prediction = actual demand observed at t - 7 days
    """
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass

    def predict(self, X_test: pd.DataFrame) -> np.array:
        """"""
        return X_test[f"rides_previous_{7*24}_hour"]

model = BaselineModelPreviousWeek()
predictions = model.predict(X_test)
predictions

test_mae = mean_absolute_error(y_test, predictions)
print(f"{test_mae=:.4f}")

## Creating baseline model 3: predicting average of last 4 weeks

class BaselineModelLast4Weeks:
    """
    Prediction = actual demand observed at t - 7 days, t - 14 days, t - 21 days, t - 28 days
    """
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass

    def predict(self, X_test: pd.DataFrame):
        # 0.25 - for the average of 4
        return 0.25*(
            X_test[f"rides_previous_{7*24}_hour"] + \
            X_test[f"rides_previous_{2*7*24}_hour"] + \
            X_test[f"rides_previous_{3*7*24}_hour"] + \
            X_test[f"rides_previous_{4*7*24}_hour"]
        )

model = BaselineModelLast4Weeks()
predictions = model.predict(X_test)
predictions

test_mae = mean_absolute_error(y_test, predictions)
print(f"{test_mae=:.4f}")
