==> notebooks/09_lightgbm_with_feature_engineering.ipynb <==
from src.paths import TRANSFORMED_DATA_DIR
import pandas as pd

df  = pd.read_parquet(TRANSFORMED_DATA_DIR / "final_data.parquet")
df

from src.data_split import train_test_split
from datetime import datetime

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2022, 8, 1, 0, 0, 0),
    target_column_name="target_rides_next_hour"
)

print(f"{X_train.shape=}")
print(f"{y_train.shape=}")
print(f"{X_test.shape=}")
print(f"{y_test.shape=}")

### New feature 1: avg_rides_last_4_weeks

def avg_rides_last_4_weeks(X: pd.DataFrame) -> pd.DataFrame:
    """
    Adds one column with the average rides from
    - 7 days ago
    - 14 days ago
    - 21 days ago
    - 28 days ago
    """
    X['avg_rides_last_4_weeks'] = 0.25*(
        X[f"rides_previous_{7*24}_hour"] + \
        X[f"rides_previous_{2*7*24}_hour"] + \
        X[f"rides_previous_{3*7*24}_hour"] + \
        X[f"rides_previous_{4*7*24}_hour"]
    )
    return X

from sklearn.preprocessing import FunctionTransformer

add_feature_avg_rides_last_4_weeks = FunctionTransformer(
    avg_rides_last_4_weeks, validate=False
)

add_feature_avg_rides_last_4_weeks.fit_transform(X_train)

### New Feature: hour of the day & days of a week form pickup_hour

from sklearn.base import BaseEstimator, TransformerMixin

class TemporalFeatureEngineering(BaseEstimator, TransformerMixin):

    def fit(self, X, y=None):
        return self
    
    def transform(self, X, y=None):
        X_ = X.copy()

        # generate numeric columns from datetime
        X_["hour"] = X_["pickup_hour"].dt.hour
        X_["day_of_week"] = X_["pickup_hour"].dt.dayofweek

        return X_.drop(columns=["pickup_hour"])

add_temporal_features = TemporalFeatureEngineering()
add_temporal_features.fit_transform(X_train)

import lightgbm as lgb

### Creating scikit-learn pipeline

from sklearn.pipeline import make_pipeline

pipeline = make_pipeline(
    add_feature_avg_rides_last_4_weeks,
    add_temporal_features,
    lgb.LGBMRegressor()
)
pipeline.fit(X_train, y_train)

predictions = pipeline.predict(X_test)
predictions

from sklearn.metrics import mean_absolute_error
test_mae = mean_absolute_error(y_test, predictions)
print(f"{test_mae=:.4f}")

from src.plot import plot_one_sample

plot_one_sample(
    features=X_test,
    targets=y_test,
    example_id=2977,
    predictions=pd.Series(predictions)
)
