==> notebooks/12_feature_pipeline.ipynb <==

from datetime import datetime, timedelta, timezone
import pandas as pd

current_date = pd.to_datetime(datetime.utcnow()).floor('H')
print(f"{current_date=}")

# fetching last 28 days data
fetch_data_to = current_date
fetch_data_from = current_date - timedelta(days=28)

print(f"{fetch_data_to=}")
print(f"{fetch_data_from=}")

### Fetching raw batch data from the original data source i.e., NYC yellow taxi data website

from src.data import load_raw_data

def fetch_batch_raw_data(from_date: datetime, to_date: datetime) -> pd.DataFrame:
    """
    Simulate production data by sampling historical data from 52 weeks ago (i.e. 1 year)
    """

    # we dont have access to recent data, so to simulate fetching data every hour, we will back date to 1 year and simulate fetching from that date

    from_date_ = from_date - timedelta(days=7*52)
    to_date_ = to_date - timedelta(days=7*52)

    # download 2 files from website
    rides = load_raw_data(year=from_date_.year, months=from_date_.month)
    rides = rides[rides.pickup_datetime >= from_date_]
    rides_2 = load_raw_data(year=to_date_.year, months=to_date_.month)
    rides_2 = rides_2[rides_2.pickup_datetime < to_date_]

    rides = pd.concat([rides, rides_2])

    # shift the data to pretend this is recent data
    rides['pickup_datetime'] += timedelta(days=7*52)

    rides.sort_values(by=['pickup_location_id', 'pickup_datetime'], inplace=True)

    return rides

rides = fetch_batch_raw_data(from_date=fetch_data_from, to_date=fetch_data_to)

### Transforming the fetched raw data to timeseries data

from src.data import transform_raw_data_into_ts_data
ts_data = transform_raw_data_into_ts_data(rides)

# string to datetime
ts_data['pickup_hour'] = pd.to_datetime(ts_data['pickup_hour'], utc=True)

# add column with Unix epoch milliseconds
ts_data['pickup_ts'] = ts_data['pickup_hour'].astype(int) // 10**6

ts_data["pickup_location_id"].dtype

ts_data["pickup_location_id"] = ts_data["pickup_location_id"].astype("int64")
ts_data["pickup_location_id"].dtype

### Make a connection to hopsworks feature store to insert this data

import hopsworks
import src.config as config

# connect to the project
project = hopsworks.login(
    project=config.HOPSWORKS_PROJECT_NAME,
    api_key_value=config.HOPSWORKS_API_KEY
)

# connect to the feature store

feature_store = project.get_feature_store()
feature_store

# connect to the feature group

feature_group = feature_store.get_or_create_feature_group(
    name=config.FEATURE_GROUP_NAME,
    version=config.FEATURE_GROUP_VERSION,
    description="Time-series data at hourly frequency",
    primary_key = ['pickup_location_id', 'pickup_ts'],
    event_time="pickup_ts",
)
feature_group

feature_group.get_feature("pickup_location_id").type

feature_group.insert(ts_data, write_options={"wait_for_job": False})
